requirements.txt
# Core Dependencies
opencv-python==4.8.1.78
numpy==1.24.3
imutils==0.5.4
Pillow==10.0.1
python-dateutil==2.8.2

# GUI Framework
tkinter

# Optional: For YOLO (more accurate detection)
ultralytics==8.0.196
torch>=1.8.0
torchvision>=0.9.0

# Additional utilities
argparse
threading
datetime